Consistent use of newlines in mini and verbose reporters
The final output of mini and verbose reporters is complex, with lots of
optional values. Consistently add newlines *after* new output is
concatenated, making it easier to reason about when to add multiple
newline characters to separate different kinds of output.

Ensure output starts and ends with but a single empty line.

Trim newlines whenever we're unsure about the line endings.

The `compare-line-output` helper now requires *all* lines are compared.
Replace the `SKIP_UNTIL_EMPTY_LINE` object with a symbol.