Reduce repetition between error message and value label
If the value label starts with the error message, don't print the
error message in the mini and verbose reporters.

Update the value labels for callback errors, thrown errors and promise
rejections in `test.js` to start with the error message. This minimizes
the output from:

```
Error thrown in test

Error:

[Error]
```

To:

```
Error thrown in test:

[Error]
```